# app/static/swagger.yaml
...
swagger: "2.0"
info:
  title: "Mechanic Shop API"
  description: "API to manage customers, mechanics, tickets, and inventory for a mechanic shop."
  version: "1.0.0"
host: "127.0.0.1:5000"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"

# MARK: Security Definitions
securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header
paths:
...
  /customers/login:
    post:
      tags: [Customers]
      summary: "Customer login"
      description: "Authenticate a customer and return a token."
      parameters:
        - in: "body"
          name: "body"
          required: true
          schema:
            $ref: "#/definitions/LoginPayload"
      responses:
        200:
          description: "Login successful"
          schema:
            $ref: "#/definitions/LoginResponse"
          examples:
            application/json:
              token: "example_token"
              message: "Login successful"
  /customers:
    post:
      tags: [Customers]
      summary: "Create new customer"
      description: "Register a new customer."
      parameters:
        - in: "body"
          name: "body"
          required: true
          schema:
            $ref: "#/definitions/CustomerCreate"
      responses:
        201:
          description: "Customer created"
          schema:
            $ref: "#/definitions/Customer"
        examples:
          application/json:
            id: 1
            name: "John Doe"
            email: "john.doe@example.com"

    get:
      tags: [Customers]
      summary: "Get all customers"
      description: "Retrieve all customers. Supports pagination."
      parameters:
        - in: "query"
          name: "page"
          type: "integer"
          required: false
        - in: "query"
          name: "per_page"
          type: "integer"
          required: false
      responses:
        200:
          description: "List of customers"
          schema:
            $ref: "#/definitions/CustomerList"
          examples:
            application/json:
              - id: 1
                name: "John Doe"
                email: "john.doe@example.com"

    put:
      tags: [Customers]
      summary: "Update customer (all fields)"
      description: "Update all fields for a customer. Requires authentication."
      security: [{ bearerAuth: [] }]
      parameters:
        - in: "body"
          name: "body"
          required: true
          schema:
            $ref: "#/definitions/CustomerUpdate"
      responses:
        200:
          description: "Customer updated"
          schema:
            $ref: "#/definitions/Customer"
          examples:
            application/json:
              id: 1
              name: "John Doe"
              email: "john.doe@example.com"

    patch:
      tags: [Customers]
      summary: "Update customer (selected fields)"
      description: "Update selected fields for a customer. Requires authentication."
      security: [{ bearerAuth: [] }]
      parameters:
        - in: "body"
          name: "body"
          required: true
          schema:
            $ref: "#/definitions/CustomerPatch"
      responses:
        200:
          description: "Customer updated"
          schema:
            $ref: "#/definitions/Customer"
          examples:
            application/json:
              id: 1
              name: "John Doe"
              email: "john.doe@example.com"

    delete:
      tags: [Customers]
      summary: "Delete customer"
      description: "Delete a customer. Requires authentication."
      security: [{ bearerAuth: [] }]
      responses:
        200:
          description: "Customer deleted"
          schema:
            $ref: "#/definitions/DeleteResponse"
          examples:
            application/json:
              message: "Customer deleted successfully"
  /customers/{id}:
    get:
      tags: [Customers]
      summary: "Get specific customer"
      description: "Retrieve details for a specific customer by ID."
      parameters:
        - in: "path"
          name: "id"
          required: true
          type: "integer"
      responses:
        200:
          description: "Customer details"
          schema:
            $ref: "#/definitions/Customer"
          examples:
            application/json:
              id: 1
              name: "John Doe"
              email: "john.doe@example.com"

  # MARK: Mechanics Endpoints
  /mechanics/login:
    post:
      tags: [Mechanics]
      summary: "Mechanic login"
      description: "Authenticate a mechanic and return a token."
      parameters:
        - in: "body"
          name: "body"
          required: true
          schema:
            $ref: "#/definitions/LoginPayload"
      responses:
        200:
          description: "Login successful"
          schema:
            $ref: "#/definitions/LoginResponse"
          examples:
            application/json:
              token: "example_token"
              message: "Login successful"
  /mechanics:
    post:
      tags: [Mechanics]
      summary: "Create mechanic"
      description: "Register a new mechanic."
      parameters:
        - in: "body"
          name: "body"
          required: true
          schema:
            $ref: "#/definitions/MechanicCreate"
      responses:
        201:
          description: "Mechanic created"
          schema:
            $ref: "#/definitions/Mechanic"
        examples:
          application/json:
            id: 1
            name: "Jane Smith"
            email: "jane.smith@example.com"

    get:
      tags: [Mechanics]
      summary: "Get all mechanics"
      description: "Retrieve all mechanics. Supports pagination."
      parameters:
        - in: "query"
          name: "page"
          type: "integer"
          required: false
        - in: "query"
          name: "per_page"
          type: "integer"
          required: false
      responses:
        200:
          description: "List of mechanics"
          schema:
            $ref: "#/definitions/MechanicList"
          examples:
            application/json:
              - id: 1
                name: "Jane Smith"
                email: "jane.smith@example.com"
  /mechanics/{id}:
    get:
      tags: [Mechanics]
      summary: "Get specific mechanic"
      description: "Retrieve details for a specific mechanic by ID."
      parameters:
        - in: "path"
          name: "id"
          required: true
          type: "integer"
      responses:
        200:
          description: "Mechanic details"
          schema:
            $ref: "#/definitions/Mechanic"
          examples:
            application/json:
              id: 1
              name: "Jane Smith"
              email: "jane.smith@example.com"

    put:
      tags: [Mechanics]
      summary: "Update mechanic (all fields)"
      description: "Update all fields for a mechanic."
      parameters:
        - in: "body"
          name: "body"
          required: true
          schema:
            $ref: "#/definitions/MechanicUpdate"
      responses:
        200:
          description: "Mechanic updated"
          schema:
            $ref: "#/definitions/Mechanic"
          examples:
            application/json:
              id: 1
              name: "Jane Smith"
              email: "jane.smith@example.com"

    patch:
      tags: [Mechanics]
      summary: "Update mechanic (selected fields)"
      description: "Update selected fields for a mechanic."
      parameters:
        - in: "body"
          name: "body"
          required: true
          schema:
            $ref: "#/definitions/MechanicPatch"
      responses:
        200:
          description: "Mechanic updated"
          schema:
            $ref: "#/definitions/Mechanic"
          examples:
            application/json: 
              id: 1
              name: "Jane Smith"
              email: "jane.smith@example.com"

    delete:
      tags: [Mechanics]
      summary: "Delete mechanic"
      description: "Delete a mechanic."
      responses:
        200:
          description: "Mechanic deleted"
          schema:
            $ref: "#/definitions/DeleteResponse"
          examples:
            application/json:
              message: "Mechanic deleted successfully"
...
  /mechanics/search:
    get:
      tags: [Mechanics]
      summary: "Search mechanics"
      description: "Search for mechanics by name."
      parameters:
        - in: "query"
          name: "name"
          type: "string"
          required: false
      responses:
        200:
          description: "Mechanics search results"
          schema:
            $ref: "#/definitions/MechanicList"
          examples:
            application/json:
              - id: 1
                name: "Jane Smith"
                email: "jane.smith@example.com"

  # MARK: Tickets Endpoints
...
  /tickets/{id}/mechanics:
    get:
      tags: [Tickets]
      summary: "Get mechanics assigned to ticket"
      description: "Retrieve all mechanics assigned to a specific ticket."
      parameters:
        - in: "path"
          name: "id"
          required: true
          type: "integer"
      responses:
        200:
          description: "Mechanics assigned to ticket"
          schema:
            $ref: "#/definitions/MechanicList"
          examples:
            application/json:
              - id: 1
                name: "Mechanic One"
                email: "mechanic.one@example.com"
              - id: 2
                name: "Mechanic Two"
                email: "mechanic.two@example.com"
...
definitions:
  LoginPayload:
    type: "object"
    properties:
      email: { type: "string" }
      password: { type: "string" }
    required: [email, password]
  LoginResponse:
    type: "object"
    properties:
      token: { type: "string" }
      message: { type: "string" }
  CustomerCreate:
    type: "object"
    properties:
      name: { type: "string" }
      email: { type: "string" }
      phone: { type: "string" }
      password: { type: "string" }
    required: [name, email, phone, password]
  CustomerUpdate:
    type: "object"
    properties:
      name: { type: "string" }
      email: { type: "string" }
      phone: { type: "string" }
      password: { type: "string" }
    required: [name, email, phone, password]
  CustomerPatch:
    type: "object"
    properties:
      name: { type: "string" }
      email: { type: "string" }
      phone: { type: "string" }
      password: { type: "string" }
  Customer:
    type: "object"
    properties:
      id: { type: "integer" }
      name: { type: "string" }
      email: { type: "string" }
      phone: { type: "string" }
  CustomerList:
    type: "array"
    items:
      $ref: "#/definitions/Customer"
  MechanicCreate:
    type: "object"
    properties:
      name: { type: "string" }
      email: { type: "string" }
      phone: { type: "string" }
      salary: { type: "number", format: "float" }
      password: { type: "string" }
    required: [name, email, phone, salary, password]
  MechanicUpdate:
    type: "object"
    properties:
      name: { type: "string" }
      email: { type: "string" }
      phone: { type: "string" }
      salary: { type: "number", format: "float" }
      password: { type: "string" }
    required: [name, email, phone, salary, password]
  MechanicPatch:
    type: "object"
    properties:
      name: { type: "string" }
      email: { type: "string" }
      phone: { type: "string" }
      salary: { type: "number", format: "float" }
      password: { type: "string" }
  Mechanic:
    type: "object"
    properties:
      id: { type: "integer" }
      name: { type: "string" }
      email: { type: "string" }
      phone: { type: "string" }
      salary: { type: "number", format: "float" }
  MechanicList:
    type: "array"
    items:
      $ref: "#/definitions/Mechanic"
...
  DeleteResponse:
    type: "object"
    properties:
      message: { type: "string" }